Created MotionBoard skeleton functions
